Update description and types for `History`

+++ src/commonMain/kotlin/entities/History.kt
@@ -10,9 +10,9 @@ import kotlin.js.JsExport
 /**
  * @property lastUploadTime The last upload time for this endpoint, in milliseconds since the UNIX epoch.
  * @property entries The historical sales.
- * @property stackSizeHistogram A [Map] of quantities to sale counts, representing the number of sales of each quantity.
- * @property stackSizeHistogramNq A [Map] of quantities to NQ sale counts, representing the number of sales of each quantity.
- * @property stackSizeHistogramHq A [Map] of quantities to HQ sale counts, representing the number of sales of each quantity.
+ * @property stackSizeHistogram A map of quantities to sale counts, representing the number of sales of each quantity.
+ * @property stackSizeHistogramNq A map of quantities to NQ sale counts, representing the number of sales of each quantity.
+ * @property stackSizeHistogramHq A map of quantities to HQ sale counts, representing the number of sales of each quantity.
  * @property regularSaleVelocity The average number of sales per day, over the past seven days (or the entirety of the shown sales, whichever comes first).
  * @property nqSaleVelocity The average number of NQ sales per day, over the past seven days (or the entirety of the shown sales, whichever comes first).
  * @property hqSaleVelocity The average number of HQ sales per day, over the past seven days (or the entirety of the shown sales, whichever comes first).
@@ -24,7 +24,7 @@ data class History(
     val stackSizeHistogram: StackSizeHistogram = null,
     @SerialName("stackSizeHistogramNQ") val stackSizeHistogramNq: StackSizeHistogram = null,
     @SerialName("stackSizeHistogramHQ") val stackSizeHistogramHq: StackSizeHistogram = null,
-    val regularSaleVelocity: Double,
-    val nqSaleVelocity: Double,
-    val hqSaleVelocity: Double,
+    val regularSaleVelocity: Float,
+    val nqSaleVelocity: Float,
+    val hqSaleVelocity: Float,
 ) : MarketBoard
